MSc/Telegram Recipe Chatbot/requirements.txt
telegram
python-telegram-bot>=13.1
configparser
redis
firebase-admin 
google-cloud-firestore
newspaper3k
sympy
Flask
pyOpenSSL